♻️ move routes back into router
